backend/requirements.txt
fastapi
uvicorn
crewai
langchain
sqlalchemy
python-dotenv
pydantic
sse-starlette
